faster testing because isolated test_ring_overlap
